Extract getItems into MiroItems
