+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/platform/transformer/transformers/miro/MiroImageApiURL.scala
@@ -1,0 +1,19 @@
+package uk.ac.wellcome.platform.transformer.transformers.miro
+
+trait MiroImageApiURL {
+  def buildImageApiURL(miroID: String, templateName: String): String = {
+    val iiifImageApiBaseUri = "https://iiif.wellcomecollection.org"
+
+    val imageUriTemplates = Map(
+      "thumbnail" -> "%s/image/%s.jpg/full/300,/0/default.jpg",
+      "info" -> "%s/image/%s.jpg/info.json"
+    )
+
+    val imageUriTemplate = imageUriTemplates.getOrElse(
+      templateName,
+      throw new Exception(
+        s"Unrecognised Image API URI template ($templateName)!"))
+
+    imageUriTemplate.format(iiifImageApiBaseUri, miroID)
+  }
+}
