+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/platform/transformer/transformers/miro/MiroItemsTest.scala
@@ -1,18 +1,20 @@
 package uk.ac.wellcome.platform.transformer.transformers.miro
 
 import org.scalatest.{FunSpec, Matchers}
+import uk.ac.wellcome.models.work.internal.DigitalLocation
 import uk.ac.wellcome.platform.transformer.source.MiroTransformableData
 
-class MiroCreditsTest extends FunSpec with Matchers {
+class MiroItemsTest extends FunSpec with Matchers{
+  val transformer = new MiroItems {}
+
   it("finds the credit line for an image-specific contributor code") {
-    transformer.getCredit(
+    transformer.getItems(
       miroId = "B0011308",
       miroData = MiroTransformableData(
         creditLine = None,
-        sourceCode = Some("FDN")
+        sourceCode = Some("FDN"),
+        useRestrictions = Some("CC-0")
       )
-    ) shouldBe Some("Ezra Feilden")
+    ).head.agent.locations.head.asInstanceOf[DigitalLocation].credit shouldBe Some("Ezra Feilden")
   }
-
-  val transformer = new MiroCredits {}
 }
